Refactoring Multi module methods to its own file. Writing method documentation.
